Add Board ID folder, add timestamp to description
